code/src/FastAPIProject/requirements.txt
openpyxl
seaborn
scikit-learn~=1.6.1
tensorflow~=2.19.0
requests~=2.32.3
python-dotenv
pandas~=2.2.3
fastapi~=0.115.12
keras~=3.9.0
numpy~=2.1.3
uvicorn
google-generativeai
secure-smtplib
jira